+++ packages/common-cosmos/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support protobuf to ts cosmos `codegen` (#160)
 
 ## [2.3.1] - 2023-07-31
 ### Changed